데이터 타입 통일, url 통일, partner image 생성, entity 이름 통일

+++ festival/src/main/java/com/example/festival/auth/AuthController.java
@@ -20,7 +20,7 @@ public class AuthController {
 
 
     @PostMapping("/join")
-    public Long join(@RequestBody User user) {
+    public Integer join(@RequestBody User user) {
         User userEntity = new User();
         userEntity.setPassword(bCryptPasswordEncoder.encode(user.getPassword())); //비밀번호 암호화
         userEntity.setIdentify(user.getIdentify()); //아이디
